Phase B: Create Models module
Extract Model type and accessors from OCP:
- Create src/Models/ manifest with AbstractModel, struct Model
- Move all Model accessors (~60 functions) to Models
- Extract time/objective/constraints accessors to Components submodules
- Fix method shadowing: dim_path_constraints_nl, dim_boundary_constraints_nl in solution.jl now properly extend Components.*
- Requalify Init/validation.jl: variable/state/control → OCP.variable/state/control
- Update OCP to use Models as dependency
- Maintain backward compatibility via reexports in CTModels.jl

All tests pass (3521/3521).

Part of OCP split refactor (Components/Models/Building/Solutions)
