+++ src/OCP/Building/model.jl
@@ -557,1106 +557,4 @@ function build_model(pre_ocp::PreModel; build_examodel=nothing)::Model
     return build(pre_ocp; build_examodel=build_examodel)
 end
 
-# ------------------------------------------------------------------------------ #
-# Getters
-# ------------------------------------------------------------------------------ #
-
-# time dependence
-"""
-$(TYPEDSIGNATURES)
-
-Return `true` for an autonomous model.
-"""
-function is_autonomous(
-    ::Model{
-        Autonomous,
-        <:TimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)
-    return true
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return `false` for a non-autonomous model.
-"""
-function is_autonomous(
-    ::Model{
-        NonAutonomous,
-        <:TimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)
-    return false
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Check whether the problem has optimisation variables.
-"""
-function is_variable(ocp::Model)::Bool
-    return variable_dimension(ocp) > 0
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Check whether the problem is control-free (no control input).
-"""
-function is_control_free(ocp::Model)::Bool
-    return control_dimension(ocp) == 0
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Check whether the problem has optimisation variables.
-
-# Arguments
-- `ocp::Model`: The optimal control problem model.
-
-# Returns
-- `Bool`: `true` if the problem has optimisation variables (variable dimension > 0), `false` otherwise.
-
-# Example
-```julia-repl
-julia> has_variable(model)  # returns true if variables are present
-```
-"""
-has_variable(ocp::Model)::Bool = is_variable(ocp)
-
-"""
-$(TYPEDSIGNATURES)
-
-Check whether the problem has control input.
-
-# Arguments
-- `ocp::Model`: The optimal control problem model.
-
-# Returns
-- `Bool`: `true` if the problem has control input (control dimension > 0), `false` otherwise.
-
-# Example
-```julia-repl
-julia> has_control(model)  # returns true if control is present
-```
-"""
-has_control(ocp::Model)::Bool = !is_control_free(ocp)
-
-"""
-$(TYPEDSIGNATURES)
-
-Check whether the problem has an abstract definition.
-
-# Arguments
-- `ocp::Model`: The optimal control problem model.
-
-# Returns
-- `Bool`: `true` if the model has a non-empty abstract definition, `false` otherwise.
-
-# Example
-```julia-repl
-julia> has_abstract_definition(model)  # returns true if definition was attached
-```
-"""
-has_abstract_definition(ocp::Model)::Bool = !__is_definition_empty(definition(ocp))
-
-"""
-$(TYPEDSIGNATURES)
-
-Check whether the problem is abstractly defined.
-
-# Arguments
-- `ocp::Model`: The optimal control problem model.
-
-# Returns
-- `Bool`: `true` if the model has a non-empty abstract definition, `false` otherwise.
-
-# Example
-```julia-repl
-julia> is_abstractly_defined(model)  # returns true if definition was attached
-```
-"""
-is_abstractly_defined(ocp::Model)::Bool = has_abstract_definition(ocp)
-
-"""
-$(TYPEDSIGNATURES)
-
-Check whether the problem is non-autonomous (time-dependent).
-
-# Arguments
-- `ocp::Model`: The optimal control problem model.
-
-# Returns
-- `Bool`: `true` if the system is non-autonomous (time-dependent), `false` otherwise.
-
-# Example
-```julia-repl
-julia> is_nonautonomous(model)  # returns true if time-dependent
-```
-"""
-is_nonautonomous(ocp::Model)::Bool = !is_autonomous(ocp)
-
-"""
-$(TYPEDSIGNATURES)
-
-Check whether the problem has no optimisation variables.
-
-# Arguments
-- `ocp::Model`: The optimal control problem model.
-
-# Returns
-- `Bool`: `true` if the problem has no optimisation variables (variable dimension == 0), `false` otherwise.
-
-# Example
-```julia-repl
-julia> is_nonvariable(model)  # returns true if no variables
-```
-"""
-is_nonvariable(ocp::Model)::Bool = !is_variable(ocp)
-
-# State
-"""
-$(TYPEDSIGNATURES)
-
-Return the state struct.
-"""
-function state(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel,
-        T,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::T where {T<:AbstractStateModel}
-    return ocp.state
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the name of the state.
-"""
-function state_name(ocp::Model)::String
-    return name(state(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the names of the components of the state.
-"""
-function state_components(ocp::Model)::Vector{String}
-    return components(state(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the state dimension.
-"""
-function state_dimension(ocp::Model)::Dimension
-    return dimension(state(ocp))
-end
-
-# Control
-"""
-$(TYPEDSIGNATURES)
-
-Return the control struct.
-"""
-function control(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel,
-        <:AbstractStateModel,
-        T,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::T where {T<:AbstractControlModel}
-    return ocp.control
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the name of the control.
-"""
-function control_name(ocp::Model)::String
-    return name(control(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the names of the components of the control.
-"""
-function control_components(ocp::Model)::Vector{String}
-    return components(control(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the control dimension.
-"""
-function control_dimension(ocp::Model)::Dimension
-    return dimension(control(ocp))
-end
-
-# Variable 
-"""
-$(TYPEDSIGNATURES)
-
-Return the variable struct.
-"""
-function variable(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        T,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::T where {T<:AbstractVariableModel}
-    return ocp.variable
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the name of the variable.
-"""
-function variable_name(ocp::Model)::String
-    return name(variable(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the names of the components of the variable.
-"""
-function variable_components(ocp::Model)::Vector{String}
-    return components(variable(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the variable dimension.
-"""
-function variable_dimension(ocp::Model)::Dimension
-    return dimension(variable(ocp))
-end
-
-# Times
-"""
-$(TYPEDSIGNATURES)
-
-Return the times struct.
-"""
-function times(
-    ocp::Model{
-        <:TimeDependence,
-        T,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::T where {T<:TimesModel}
-    return ocp.times
-end
-
-# Time name
-"""
-$(TYPEDSIGNATURES)
-
-Return the name of the time.
-"""
-function time_name(ocp::Model)::String
-    return time_name(times(ocp))
-end
-
-# Initial time
-"""
-$(TYPEDSIGNATURES)
-
-Throw an error for unsupported initial time access.
-"""
-function initial_time(ocp::AbstractModel)
-    throw(
-        Exceptions.PreconditionError(
-            "Cannot get initial time with this function";
-            reason="This model type does not support direct initial time access",
-            suggestion="Use initial_time(ocp) on a Model with FixedTimeModel or use initial_time(ocp, variable) for variable initial time",
-            context="initial_time on AbstractModel",
-        ),
-    )
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Throw an error for unsupported initial time access with variable.
-"""
-function initial_time(ocp::AbstractModel, variable::AbstractVector)
-    throw(
-        Exceptions.PreconditionError(
-            "Cannot get initial time with this function";
-            reason="This model type does not support initial time access with variable",
-            suggestion="Ensure the model has variable initial time configured, or use initial_time(ocp) for fixed initial time",
-            context="initial_time with variable on AbstractModel",
-        ),
-    )
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the initial time, for a fixed initial time.
-"""
-function initial_time(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel{FixedTimeModel{T},<:AbstractTimeModel},
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::T where {T<:Time}
-    return initial_time(times(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the initial time, for a free initial time.
-"""
-function initial_time(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel{FreeTimeModel,<:AbstractTimeModel},
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-    variable::AbstractVector{T},
-)::T where {T<:ctNumber}
-    return initial_time(times(ocp), variable)
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the initial time, for a free initial time.
-"""
-function initial_time(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel{FreeTimeModel,<:AbstractTimeModel},
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-    variable::T,
-)::T where {T<:ctNumber}
-    return initial_time(times(ocp), [variable])
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the name of the initial time.
-"""
-function initial_time_name(ocp::Model)::String
-    return initial_time_name(times(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Check if the initial time is fixed.
-"""
-function has_fixed_initial_time(ocp::Model)::Bool
-    return has_fixed_initial_time(times(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Check if the initial time is free.
-"""
-function has_free_initial_time(ocp::Model)::Bool
-    return has_free_initial_time(times(ocp))
-end
-
-# Final time
-"""
-$(TYPEDSIGNATURES)
-
-Throw an error for unsupported final time access.
-"""
-function final_time(ocp::AbstractModel)
-    throw(
-        Exceptions.PreconditionError(
-            "Cannot get final time with this function";
-            reason="This model type does not support direct final time access",
-            suggestion="Use final_time(ocp) on a Model with FixedTimeModel or use final_time(ocp, variable) for variable final time",
-            context="final_time on AbstractModel",
-        ),
-    )
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Throw an error for unsupported final time access with variable.
-"""
-function final_time(ocp::AbstractModel, variable::AbstractVector)
-    throw(
-        Exceptions.PreconditionError(
-            "Cannot get final time with this function";
-            reason="This model type does not support final time access with variable",
-            suggestion="Ensure the model has variable final time configured, or use final_time(ocp) for fixed final time",
-            context="final_time with variable on AbstractModel",
-        ),
-    )
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the final time, for a fixed final time.
-"""
-function final_time(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel{<:AbstractTimeModel,FixedTimeModel{T}},
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::T where {T<:Time}
-    return final_time(times(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the final time, for a free final time.
-"""
-function final_time(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel{<:AbstractTimeModel,FreeTimeModel},
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-    variable::AbstractVector{T},
-)::T where {T<:ctNumber}
-    return final_time(times(ocp), variable)
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the final time, for a free final time.
-"""
-function final_time(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel{<:AbstractTimeModel,FreeTimeModel},
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-    variable::T,
-)::T where {T<:ctNumber}
-    return final_time(times(ocp), [variable])
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the name of the final time.
-"""
-function final_time_name(ocp::Model)::String
-    return final_time_name(times(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Check if the final time is fixed.
-"""
-function has_fixed_final_time(ocp::Model)::Bool
-    return has_fixed_final_time(times(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Check if the final time is free.
-"""
-function has_free_final_time(ocp::Model)::Bool
-    return has_free_final_time(times(ocp))
-end
-
-# Objective
-"""
-$(TYPEDSIGNATURES)
-
-Return the objective struct.
-"""
-function objective(
-    ocp::Model{
-        <:TimeDependence,
-        <:AbstractTimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        O,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::O where {O<:AbstractObjectiveModel}
-    return ocp.objective
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the type of criterion (:min or :max).
-"""
-function criterion(ocp::Model)::Symbol
-    return criterion(objective(ocp))
-end
-
-# Mayer
-"""
-$(TYPEDSIGNATURES)
-
-Throw an error when accessing Mayer cost on a model without one.
-"""
-function mayer(ocp::AbstractModel)
-    throw(
-        Exceptions.PreconditionError(
-            "Cannot access Mayer cost";
-            reason="This OCP has no Mayer objective defined",
-            suggestion="Define a Mayer objective using objective!(ocp, :min/:max, mayer=...) before accessing it",
-            context="mayer accessor",
-        ),
-    )
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the Mayer cost.
-"""
-function mayer(
-    ocp::Model{
-        <:TimeDependence,
-        <:AbstractTimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:MayerObjectiveModel{M},
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::M where {M<:Function}
-    return mayer(objective(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the Mayer cost.
-"""
-function mayer(
-    ocp::Model{
-        <:TimeDependence,
-        <:AbstractTimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:BolzaObjectiveModel{M,<:Function},
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::M where {M<:Function}
-    return mayer(objective(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Check if the model has a Mayer cost.
-"""
-function has_mayer_cost(ocp::Model)::Bool
-    return has_mayer_cost(objective(ocp))
-end
-
-# Lagrange
-"""
-$(TYPEDSIGNATURES)
-
-Throw an error when accessing Lagrange cost on a model without one.
-"""
-function lagrange(ocp::AbstractModel)
-    throw(
-        Exceptions.PreconditionError(
-            "Cannot access Lagrange cost";
-            reason="This OCP has no Lagrange objective defined",
-            suggestion="Define a Lagrange objective using objective!(ocp, :min/:max, lagrange=...) before accessing it",
-            context="lagrange accessor",
-        ),
-    )
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the Lagrange cost.
-"""
-function lagrange(
-    ocp::Model{
-        <:TimeDependence,
-        <:AbstractTimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        LagrangeObjectiveModel{L},
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::L where {L<:Function}
-    return lagrange(objective(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the Lagrange cost.
-"""
-function lagrange(
-    ocp::Model{
-        <:TimeDependence,
-        <:AbstractTimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:BolzaObjectiveModel{<:Function,L},
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::L where {L<:Function}
-    return lagrange(objective(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Check if the model has a Lagrange cost.
-"""
-function has_lagrange_cost(ocp::Model)::Bool
-    return has_lagrange_cost(objective(ocp))
-end
-
-# Dynamics
-"""
-$(TYPEDSIGNATURES)
-
-Return the dynamics.
-"""
-function dynamics(
-    ocp::Model{
-        <:TimeDependence,
-        <:AbstractTimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        D,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::D where {D<:Function}
-    return ocp.dynamics
-end
-
-# build_examodel
-"""
-$(TYPEDSIGNATURES)
-
-Return the build_examodel.
-"""
-function get_build_examodel(
-    ocp::Model{
-        <:TimeDependence,
-        <:AbstractTimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        BE,
-    },
-)::BE where {BE<:Function}
-    return ocp.build_examodel
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Fallback method: throws a `PreconditionError` explaining that the `:exa` modeler is not available because the `Model` was assembled through the functional (macro-free) API and therefore carries no Exa builder.
-"""
-function get_build_examodel(
-    ::Model{
-        <:TimeDependence,
-        <:AbstractTimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        <:AbstractDefinition,
-        <:Nothing,
-    },
-)
-    throw(
-        Exceptions.PreconditionError(
-            "The :exa modeler is not available for this model";
-            reason="this Model was built with the functional (macro-free) API (PreModel + time!/state!/control!/variable!/dynamics!/objective!/constraint! + build), which does not generate the Exa builder required by the Exa (:exa) modeler",
-            suggestion="either choose another modeler, e.g. ADNLP (:adnlp), or define the optimal control problem with the @def macro so that the Exa builder is generated",
-            context="get_build_examodel called on a Model built without an Exa builder",
-        ),
-    )
-end
-
-# Constraints
-"""
-$(TYPEDSIGNATURES)
-
-Return the constraints struct.
-"""
-function constraints(
-    ocp::Model{
-        <:TimeDependence,
-        <:AbstractTimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        C,
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::C where {C<:AbstractConstraintsModel}
-    return ocp.constraints
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return true if the model has constraints or false if not.
-"""
-function isempty_constraints(ocp::Model)::Bool
-    return Base.isempty(constraints(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the nonlinear path constraints.
-"""
-function path_constraints_nl(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:ConstraintsModel{TP,<:Tuple,<:Tuple,<:Tuple,<:Tuple},
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::TP where {TP<:Tuple}
-    return constraints(ocp).path_nl
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the nonlinear boundary constraints.
-"""
-function boundary_constraints_nl(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:ConstraintsModel{<:Tuple,TB,<:Tuple,<:Tuple,<:Tuple},
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::TB where {TB<:Tuple}
-    return constraints(ocp).boundary_nl
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the box constraints on state.
-"""
-function state_constraints_box(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:ConstraintsModel{<:Tuple,<:Tuple,TS,<:Tuple,<:Tuple},
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::TS where {TS<:Tuple}
-    return constraints(ocp).state_box
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the box constraints on control.
-"""
-function control_constraints_box(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:ConstraintsModel{<:Tuple,<:Tuple,<:Tuple,TC,<:Tuple},
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::TC where {TC<:Tuple}
-    return constraints(ocp).control_box
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the box constraints on variable.
-"""
-function variable_constraints_box(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:ConstraintsModel{<:Tuple,<:Tuple,<:Tuple,<:Tuple,TV},
-        <:AbstractDefinition,
-        <:Union{Function,Nothing},
-    },
-)::TV where {TV<:Tuple}
-    return constraints(ocp).variable_box
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the dimension of nonlinear path constraints.
-"""
-function dim_path_constraints_nl(ocp::Model)::Dimension
-    return dim_path_constraints_nl(constraints(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the dimension of the boundary constraints.
-"""
-function dim_boundary_constraints_nl(ocp::Model)::Dimension
-    return dim_boundary_constraints_nl(constraints(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the dimension of box constraints on state.
-"""
-function dim_state_constraints_box(ocp::Model)::Dimension
-    return dim_state_constraints_box(constraints(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the dimension of box constraints on control.
-"""
-function dim_control_constraints_box(ocp::Model)::Dimension
-    return dim_control_constraints_box(constraints(ocp))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the dimension of box constraints on variable.
-"""
-function dim_variable_constraints_box(ocp::Model)::Dimension
-    return dim_variable_constraints_box(constraints(ocp))
-end
-
-# ------------------------------------------------------------------------------ #
-# Definition getters
-# ------------------------------------------------------------------------------ #
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the model definition of the optimal control problem.
-
-# Arguments
-
-- `ocp::Model`: The built optimal control problem model.
-
-# Returns
-
-- `AbstractDefinition`: The [`Definition`](@ref) wrapping the symbolic
-  expression, or an [`EmptyDefinition`](@ref) if the user did not attach one
-  before [`build`](@ref).
-"""
-function definition(
-    ocp::Model{
-        <:TimeDependence,
-        <:TimesModel,
-        <:AbstractStateModel,
-        <:AbstractControlModel,
-        <:AbstractVariableModel,
-        <:Function,
-        <:AbstractObjectiveModel,
-        <:AbstractConstraintsModel,
-        D,
-        <:Union{Function,Nothing},
-    },
-)::D where {D<:AbstractDefinition}
-    return ocp.definition
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the symbolic expression of the model definition of a built optimal control
-problem.
-
-Delegates to [`expression`](@ref) on the underlying [`AbstractDefinition`](@ref):
-returns `d.expr` if a [`Definition`](@ref) was attached before [`build`](@ref),
-or `:(begin end)` if the definition is an [`EmptyDefinition`](@ref).
-
-# Arguments
-
-- `ocp::Model`: The built optimal control problem model.
-
-# Returns
-
-- `Expr`: The symbolic expression, or `:(begin end)` if no definition was attached.
-"""
-Components.expression(ocp::Model)::Expr = Components.expression(definition(ocp))
+# Model accessor methods are now in src/Models/model.jl.
