+++ src/OCP/OCP.jl
@@ -39,7 +39,10 @@ using MLStyle: MLStyle
 # Components provides all foundational types and type aliases
 using ..Components
 
-# Load types (Model, PreModel — Component types come from ..Components)
+# Models provides AbstractModel, Model and their accessors
+using ..Models
+
+# Load types (PreModel — Model/AbstractModel come from ..Models)
 include(joinpath(@__DIR__, "Types", "model.jl"))
 include(joinpath(@__DIR__, "Types", "solution.jl"))
 
@@ -67,8 +70,8 @@ include(joinpath(@__DIR__, "Building", "interpolation_helpers.jl"))
 include(joinpath(@__DIR__, "Building", "model.jl"))
 include(joinpath(@__DIR__, "Building", "solution.jl"))
 
-# Types defined in OCP (Component types come from ..Components — no re-export needed)
-export Model, PreModel, AbstractModel
+# Types defined in OCP (Model/AbstractModel come from ..Models; Component types from ..Components)
+export PreModel
 export Solution, AbstractSolution
 export DualModel, AbstractDualModel
 export SolverInfos, AbstractSolverInfos
@@ -82,42 +85,20 @@ export build_solution, build, build_model
 export definition!, time_dependence!
 export append_box_constraints!
 
-# Accessors — time and model structure
-export constraint, constraints
-export initial_time, final_time, time_name, time_grid, times
-export initial_time_name, final_time_name
+# Accessors — solution and solver (Models/Components functions are extended, not re-exported)
+export time_grid
 export clean_component_symbols, time_grid_model
-export criterion, has_mayer_cost, has_lagrange_cost
-export is_mayer_cost_defined, is_lagrange_cost_defined
-export has_fixed_initial_time, has_free_initial_time
-export has_fixed_final_time, has_free_final_time
-export is_autonomous
-export is_variable, is_control_free
-export has_variable, has_control
-export has_abstract_definition, is_abstractly_defined
-export is_nonautonomous, is_nonvariable
-export is_initial_time_fixed, is_initial_time_free
-export is_final_time_fixed, is_final_time_free
-export state_dimension, control_dimension, variable_dimension
-export state_name, control_name, variable_name
-export state_components, control_components, variable_components
 export control_interpolation
-# Constraint accessors
-export path_constraints_nl, boundary_constraints_nl
-export state_constraints_box, control_constraints_box, variable_constraints_box
-export dim_path_constraints_nl, dim_boundary_constraints_nl
-export dim_state_constraints_box, dim_control_constraints_box, dim_variable_constraints_box
 export dim_dual_state_constraints_box,
     dim_dual_control_constraints_box, dim_dual_variable_constraints_box
-export state, control, variable, costate, objective
-export dynamics, mayer, lagrange
-export definition, dual
+export costate
+export dual
 export iterations, status, message, success, successful
 export constraints_violation, infos
-export get_build_examodel
 export is_empty, is_empty_time_grid
 export index
 export model
+# State/control/variable solution accessors (extend Models functions — no re-export)
 # Dual constraints accessors
 export path_constraints_dual, boundary_constraints_dual
 export state_constraints_lb_dual, state_constraints_ub_dual
